backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
anthropic==0.18.0
python-dotenv==1.0.0
pydantic==2.5.0